macos Cursor, DIsplay bug fix
